\name{genefilter-deprecated}
\alias{genefilter-deprecated}
\title{Deprecated functions in package \sQuote{genefilter}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{genefilter} only, and will be defunct at the next release.
}

\details{
  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{eSetFilter}
    \item{getFilterNames}
    \item{getFuncDesc}
    \item{getRdAsText}
    \item{parseDesc}
    \item{parseArgs}
    \item{showESet}
    \item{setESetArgs}
    \item{isESet}
  }
}
