\name{tdata}
\alias{tdata}
\non_function{}
\title{A small test dataset of Affymetrix Expression data. }
\usage{data(tdata)}
\description{
The \code{tdata} data frame has 500 rows and 26 columns.
The columns correspond to samples while the rows correspond to genes.
The row names are Affymetrix accession numbers.
}
\format{
  This data frame contains 26 columns.
}

\source{
 An unknown data set.
}

\examples{
data(tdata)
}
\keyword{datasets}
